refactor : 파일 경로 변경 및 alias 경로 수정
